Break up JaiaContext
